reorg test_accessor.py + add comments
